Created individual components from Profile.js
